% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data_args}
\alias{check_data_args}
\title{Checks whether one and only one of the three arguments that can be used to pass data was specified.}
\usage{
check_data_args(df = NULL, csv = NULL, svc = FALSE)
}
\arguments{
\item{df}{Data frame that contains the data which is to be examined.}

\item{csv}{Character that specifies the path to the csv file that is to be read.}

\item{svc}{Logical that indicates whether the data shall be parsed using ruODK's \code{\link[ruODK]{odata_submission_get}}.}
}
\value{
Data frame
}
\description{
To do so, the function checks whether the args are NULL which is also there default.
If more or less than are not NULL, the function throws an error.
If exactly one argument is specified, the functions reads the data it contains and returns it in a data frame.
}
