% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_instanceIDs.R
\name{find_missing_instanceIDs}
\alias{find_missing_instanceIDs}
\title{Find all missing submissions}
\usage{
find_missing_instanceIDs(csv = NULL, df = NULL, id_col)
}
\arguments{
\item{csv}{Character that specifies the path to the csv file that is to be read. (Either csv or df must not null)}

\item{df}{Data frame that, specifies the data frame that is to be read. (Either csv or df must be null)}

\item{id_col}{Character that specifies the exact name of the instance ID in the df/csv.}
}
\value{
List
}
\description{
This function finds all instance ID's of submissions of an ODK form which are already stored on ODK Central but not loaded in your current data.
It does so by downloading the most recent submission list from ODK Central for the respective form and compares the instance IDs with the ones
that are contained in the given data.
}
\examples{
\dontrun{
# ruODK needs to be set up for this function to work
repvisforODK::setup_ruODK(svc = example/svc.svc, un = exampleusername, pw = examplepassword, tz = 'Europe/Berlin', verbose = TRUE)

find_missing_instanceIDs(df = df_odk_data, id_col = 'instance_id')
}
}
