% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_title}
\alias{fit_title}
\title{Splits a string after certain amount of characters, finds the last white space and inserts an html line break (<br>).}
\usage{
fit_title(string, n)
}
\arguments{
\item{string}{Character that contains the text which is to be split.}

\item{n}{Specifies the number of characters before which the last word is looked for.}
}
\value{
Character
}
\description{
Splits a string after certain amount of characters, finds the last white space and inserts an html line break (<br>).
}
