% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_instances.R
\name{get_new_instances}
\alias{get_new_instances}
\title{Download all missing instances}
\usage{
get_new_instances(id_col, csv = NULL, df = NULL, merge_data = TRUE)
}
\arguments{
\item{id_col}{Character that specifies the exact name of the instance ID in the df/csv.}

\item{csv}{Character that specifies the path to the csv file that is to be read.}

\item{df}{Data frame that specifies the data frame that is to be read.}

\item{merge_data}{Boolean that specifies whether the new data shall be merged with the one that was given or not.}
}
\value{
Data frame
}
\description{
Find and download all instances of an ODK form which are already stored on ODK Central but not loaded in your current data
and optionally append them to the data that was passed.
}
\examples{
get_new_instances('meta-instanceID', df=df_senegal)
}
