% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_instances_odata.R
\name{get_new_instances_odata}
\alias{get_new_instances_odata}
\title{Download all missing instances through ODATA}
\usage{
get_new_instances_odata(
  id_col,
  submission_date_col,
  csv = NULL,
  df = NULL,
  merge_data = TRUE,
  force_timezone = TRUE
)
}
\arguments{
\item{id_col}{Character that specifies the exact name of the instance ID column in the df/csv.}

\item{submission_date_col}{Character that specifies the exact name of the submission date column in the df/csv.}

\item{csv}{Character that specifies the path to the csv file that is to be read. (Either csv or df must not null)}

\item{df}{Data frame that, specifies the data frame that is to be read. (Either csv or df must be null)}

\item{merge_data}{Boolean that specifies whether the new data shall be merged with the one that was given or not.}

\item{force_timezone}{If TRUE all time stamp values will be remain the same but their timezone will be changed to the one used in the old data. If FALSE time stamps will be converted in accordance with the timezone of the old data.}
}
\value{
Data frame
}
\description{
Just like the get_new_instances() function, the get_new_instances_odata() function finds and downloads all submissions
which are already stored on ODK Central but not contained in the passed data. In contrast to its sister function, get_new_instances_odata()
does not create a delta of instance ID's between old and new data to find the missing instances but it identifies the latest
submission date in the  old data and uses it to create ODATA query that filters for submissions that were either submitted or updated or both
after this date. The new submissions can be returned by themselves or merged with the old data.
}
\examples{
*Placeholder*
}
