% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_submissions.R
\name{get_new_submissions}
\alias{get_new_submissions}
\title{Download all missing submissions}
\usage{
get_new_submissions(id_col, csv = NULL, df = NULL, merge_data = TRUE)
}
\arguments{
\item{id_col}{Character that specifies the exact name of the instance ID in the df/csv.}

\item{csv}{Character that specifies the path to the csv file that is to be read.}

\item{df}{Data frame that specifies the data frame that is to be read.}

\item{merge_data}{Boolean that specifies whether the new data shall be merged with the one that was given or not.}
}
\value{
Data frame
}
\description{
This function uses the \code{\link{find_missing_instanceIDs}} function to find all submissions of an ODK form which are already stored on ODK Central but not loaded in your current data submissions and then downloads them.
The new submissions can either be appended to the old data or be returned separately. To do so, the function makes use of ruODK's \code{\link[ruODK]{submission_get}} function which
sends GET requests to ODK Centrals REST-API to retrieve data.
}
