% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_choice_questions.R
\name{identify_choice_questions}
\alias{identify_choice_questions}
\title{Identifies all questions in a form that have choices.}
\usage{
identify_choice_questions(
  lang = NULL,
  df_schema_ext = NULL,
  choice_col = NULL,
  label_col = NULL
)
}
\arguments{
\item{lang}{Character containing the name of the language that is to be examined, defaults to NULL.}

\item{df_schema_ext}{Data frame that defines the schema of the from. Can be passed to the function to avoid downloading it multiple times. Optional, defaults to NULL.}

\item{choice_col}{String specifying the choices column that is to be changed, defaults to NULL.}

\item{label_col}{String specifying the labels column that is to be changed, defaults to NULL.}
}
\value{
List
}
\description{
The function also accounts for different languages when the lang argument is specified.
In both cases the relevant question label and choice column is named the same to facilitate code progression.
}
\examples{
\dontrun{
# 1. Without df_schema_ext
choice_questions <- identify_choice_questions(lang = 'english', choice_col = 'choices_english_(en)', label_col = 'label_english_(en)')

# 2. With df_schema_ext
# ruODK needs to be set up for this function to work
repvisforODK::setup_ruODK(svc = 'example/svc.svc', un = 'exampleusername', pw = 'examplepassword', tz = 'Europe/Berlin', verbose = TRUE)

df_schema <- ruODK::form_schema_ext()

choice_questions <- identify_choice_questions(lang = 'english', df_schema_ext = df_schema, choice_col = 'choices_english_(en)', label_col = 'label_english_(en)')
}
}
