% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_msg}
\alias{info_msg}
\title{Outputs message in CLI with an info symbol in magenta color.}
\usage{
info_msg(msg)
}
\arguments{
\item{msg}{String that is used as the message content.}
}
\description{
Outputs message in CLI with an info symbol in magenta color.
}
