% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data_sub_date}
\alias{load_data_sub_date}
\title{Downloads data from ODK Central with formatted submission date column.}
\usage{
load_data_sub_date(tz)
}
\arguments{
\item{tz}{String that defines the timezone.}
}
\value{
data frame
}
\description{
This function is necessary because this operation couldn't be performed in shiny.
}
