% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_html_tags}
\alias{remove_html_tags}
\title{Removes all HTML tags from a string.}
\usage{
remove_html_tags(html_string)
}
\arguments{
\item{html_string}{}
}
\value{
Character
}
\description{
To do so, the function uses a Regular Expression which looks for any instances with the pattern '<.*?>'.
}
