% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_choice_questions.R
\name{rename_schema}
\alias{rename_schema}
\title{Changes the name of the labels and choices column in an ODK form schema.}
\usage{
rename_schema(df, lang = NULL, choice_col = NULL, label_col = NULL)
}
\arguments{
\item{df}{Data frame containing the ODK form schema}

\item{lang}{String specifying the language of the choice/label column which is to be changed. Optional, defaults to NULL.}

\item{choice_col}{String specifying the choices column that is to be changed, defaults to NULL.}

\item{label_col}{String specifying the labels column that is to be changed, defaults to NULL.}
}
\value{
Data frame
}
\description{
This function to give the column used in another function a uniform name so that no matter what language (or no language at all) is chosen,
all functions still work as they refer to the uniform name given by this function.
}
