% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all_plots.R
\name{render_all_plots}
\alias{render_all_plots}
\title{Renders a markdown file that contains all plots of the package and creates an html file.}
\usage{
render_all_plots(
  output_file,
  title,
  csv,
  date_col,
  daily_submission_goal = 0,
  exclude_weekend = TRUE
)
}
\arguments{
\item{output_file}{String that specifies the path where the generated file shall be stored.}

\item{title}{String that specifies the title parameter in the YAML header of the document.}

\item{csv}{String that specifies the file path to the csv file that contians the ODK data.}

\item{date_col}{String that specifies the date or time stamp column in the data which is to be examined.}

\item{daily_submission_goal}{Integer or float that defines the number of daily submissions goal.}

\item{exclude_weekend}{Logical that determines whether weekends are excluded in the plot. Optional, defaults to TRUE.}
}
\value{

}
\description{
The rendered file is based on the all_plots.rmd template that comes with the package. It contains all plots that the package offers.
If you want to customize the markdown file before rendering it, you can access the template using \code{\link[rmarkdown]{draft}}.
For the latter, use the following arguments: file = 'INSERT_FILE_NAME', template = 'all_plots', package = 'repvisforODK, edit = TRUE'.
}
