% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server code for the shiny app that users can use to do generate reports.}
\usage{
server(input, output)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Server code for the shiny app that users can use to do generate reports.
}
