% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_color}
\alias{set_color}
\title{Outputs a string containing one or vector containing multiple color codes.}
\usage{
set_color(val)
}
\arguments{
\item{val}{Character that specifies which color code is shall be returned.}
}
\value{
Character character vector
}
\description{
The function is supposed to help in keeping a concise color schema.
}
