% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_ruODK}
\alias{setup_ruODK}
\title{Sets up ruODK}
\usage{
setup_ruODK(
  svc,
  un = Sys.getenv("ODKC_UN"),
  pw = Sys.getenv("ODKC_PW"),
  tz = Sys.getenv("ODKC_TZ"),
  verbose = TRUE
)
}
\arguments{
\item{svc}{Character}

\item{un}{Optional, character}

\item{pw}{Optional, character}

\item{tz}{Optional, character}

\item{verbose}{Optional, logical}
}
\value{
None
}
\description{
Establishes connection with ODK Cnetral servers by passing log in credentials, timezone and svc to the
ruODK::ru_setup() function. By default it tries to retrieve this information from local system variables.
}
\examples{
setup_ruODK('https://research.odk.path.org/v1/projects/4/forms/test.svc')
}
