% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_timeseries_lineplot.R
\name{submissions_timeseries_lineplot}
\alias{submissions_timeseries_lineplot}
\title{Generating a line plot that shows the number of submissions over time}
\usage{
submissions_timeseries_lineplot(
  df,
  date_col,
  daily_submission_goal = 0,
  exclude_weekend = TRUE,
  cumulative = TRUE
)
}
\arguments{
\item{df}{Data frame that contains the data which is to be examined.}

\item{date_col}{String that specifies the date or time stamp column in the data which is to be examined.}

\item{daily_submission_goal}{Integer or float that defines the number of daily submissions goal.}

\item{exclude_weekend}{Logical that determines whether weekends are excluded in the plot. Optional, defaults to TRUE.}

\item{cumulative}{Logical that determines whether the cumulative sum of submissions is used as values for y. Optional, defaults to TRUE.}
}
\value{
Plotly html-widget
}
\description{
The function shows a line plot with the number of submissions per day. It does account for days where no submissions were recorded by adding
the missing days wit a value of 0 so that also days were no data was collected are represented in the plot.
Optionally, a vertical line can be added that represents the daily number of submissions goal.
Further, weekends can be excluded since often there is no data collection happening and would therefore make the plot less concise and comprehensive.
Below the plot, there is a range slider which can be used to define the range of time that the plot shows. Depending on the overall time span of the
data, there are several buttons (1w = 1 week, 1m = 1 month, 6m = ..., 1y = ..., or all) that can be used to pre-define a time window that is applied to the range slider.
In order to determine the collection period the function \code{\link{collection_period}} is used.
}
