% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_period.R
\name{collection_period}
\alias{collection_period}
\title{Get the earliest and latest of your data}
\usage{
collection_period(df)
}
\arguments{
\item{df}{Data frame containing the ODK data that is to be used.}
}
\value{
list
}
\description{
The function finds the earliest and the latest submission dates in a data frame. It accounts for different spellings of the submission date column
by applying a case-insensitive RegEx formula that matches every combination of the words 'submission' and 'date'.
}
