% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_instances.R
\name{find_missing_instanceIDs}
\alias{find_missing_instanceIDs}
\title{Find all missing instances}
\usage{
find_missing_instanceIDs(id_col, csv = NULL, df = NULL)
}
\arguments{
\item{id_col}{Character that specifies the exact name of the instance ID in the df/csv.}

\item{csv}{Character that specifies the path to the csv file that is to be read.}

\item{df}{Data frame that, specifies the data frame that is to be read.}
}
\value{
Character vector
}
\description{
Find all instance ID's of an ODK form which are already stored on ODK Central but not loaded in your current data.
}
\examples{
find_missing_instanceIDs('https://research.odk.path.org/v1/projects/4/forms/02-TIMCI-SPA-cgei.svc', 'meta-instanceID', df=df_test)
}
