% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_instanceIDs.R
\name{find_missing_instanceIDs}
\alias{find_missing_instanceIDs}
\title{Find all missing submissions}
\usage{
find_missing_instanceIDs(id_col, csv = NULL, df = NULL)
}
\arguments{
\item{id_col}{Character that specifies the exact name of the instance ID in the df/csv.}

\item{csv}{Character that specifies the path to the csv file that is to be read.}

\item{df}{Data frame that, specifies the data frame that is to be read.}
}
\value{
List
}
\description{
This function finds all instance ID's of submissions of an ODK form which are already stored on ODK Central but not loaded in your current data.
It does so by downloading the most recent submission list from ODK Central for the respective form and compares the instance IDs with the ones
that are contained in the given data.
}
