% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_text_wordcloud.R
\name{free_text_wordcloud}
\alias{free_text_wordcloud}
\title{Generates a wordcloud based on text from free text questions in an ODK form.}
\usage{
free_text_wordcloud(
  svc = TRUE,
  df = NULL,
  csv = NULL,
  text_col,
  lang_wc,
  lang = "english"
)
}
\arguments{
\item{svc}{Logical that indicates whether the data shall be parsed using ruODK's \code{\link[ruODK]{odata_submission_get}}. Optional, defaults to FALSE.}

\item{df}{Data frame containing the ODK data that is to be used. Optional, defaults to NULL.}

\item{csv}{Character that specifies the path to the csv file that is to be read. Optional, defaults to NULL.}

\item{text_col}{Character or Character vector (if multiple questions shall be examined) that specifies the names of the columns of the free text questions.}

\item{lang_wc}{Character that specifies the language of the answers of the free text question. Check \code{\link[tm]{stopwords}} to find out more about stopwords list options.}

\item{lang}{Character that specifies the language of the answers of the free text question. Check \code{\link[tm]{stopwords}} to find out more about stopwords list options.}
}
\value{
List
}
\description{
The main computation, namely the creation of the corpus, the preprocessing of the text in the corpus
and the generation of the wordcloud is done through the function \code{\link{preprocess_wc_generation}}.
This function serves as a wrapper that applies the \code{\link{collection_period}} to every free text question
column that the user specified in the argument text_col.
}
