% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_calendar.R
\name{heatmap_calendar}
\alias{heatmap_calendar}
\title{Generate a calendar heat map that shows the number of submissions for each day of the data collection period.}
\usage{
heatmap_calendar(df, date_col)
}
\arguments{
\item{df}{Data frame that contains the data which is to be examined.}

\item{date_col}{String that specifies the date or time stamp column in the data which is to be examined.}
}
\value{
This function returns a ggplot object which contains a calendar heat map.
}
\description{
The heat map can be created using any date or time stamp feature.
}
