% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_wday_hourofday.R
\name{heatmap_wday_hourofday}
\alias{heatmap_wday_hourofday}
\title{Generates a heat map of the number of submissions per weekday and hour.}
\usage{
heatmap_wday_hourofday(df, date_col)
}
\arguments{
\item{df}{Data frame that contains the data which is to be examined.}

\item{date_col}{String that specifies the date or time stamp column in the data which is to be examined.}
}
\value{
Plotly html-widget
}
\description{
Takes any timestamp column in the data and aggregates by weekday and hour of the day. The generated plot then shows for each combination
the frequency.
}
