% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_choice_questions.R
\name{identify_choice_questions}
\alias{identify_choice_questions}
\title{Identifies all questions in a form that have choices.}
\usage{
identify_choice_questions(lang = NULL, df_schema_ext = NULL)
}
\arguments{
\item{lang}{Character containing the name of the language that is to be examined, defaults to NULL.}

\item{df_schema_ext}{Data frame that defines the schema of the from. Can be passed to the function to avoid downloading it multiple times. Optional, defaults to NULL.}
}
\value{
List
}
\description{
The function also accounts for different languages when the lang argument is specified.
In both cases the relevant question label and choice column is named the same to facilitate code progression.
}
