% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_choice_question_bar.R
\name{multiple_choice_question_bar}
\alias{multiple_choice_question_bar}
\title{Generates a bar plot displaying the percentage of surveyees who selected a choice in a multiple choice questions out of all that answered the question.}
\usage{
multiple_choice_question_bar(
  svc = TRUE,
  df = NULL,
  csv = NULL,
  qvec = NULL,
  lang = NULL,
  delimiter = " "
)
}
\arguments{
\item{svc}{Logical that indicates whether the data shall be parsed using ruODK's \code{\link[ruODK]{odata_submission_get}}. Optional, defaults to FALSE.}

\item{df}{Data frame containing the ODK data that is to be used. Optional, defaults to NULL.}

\item{csv}{Character that specifies the path to the csv file that is to be read. Optional, defaults to NULL.}

\item{qvec}{Character vector containing the column names of the single-choice questions that is to be examined.}

\item{lang}{Character containing the name of the language that is to be examined.}

\item{delimiter}{Character specifying the symbol that is used to separate multiple choices in the data. Optional, defaults to ' '.}
}
\value{
List
}
\description{
This function generally differentiates between parsing data directly from ODK using the svc option or passing data locally (df or csv).
If one provides an svc, the function automatically finds all questions that received multiple choices as answers. Please note, that
questions which allow multiple choices but only have received single choices are not identified as such and thus will not be displayed.
This is because the extended from schema which is provided by the ODK Central API and accessed using ruODK's \code{\link[ruODK]{form_schema_ext}} does not provide this info.
After identifying the questions, the functions also maps the question and choices labels according to their names so that the final plot
contains a meaningful title and legend. By default, the function assumes that there is no language differentiation implemented in the form.
If so, the argument lang has to be specified accordingly (see parameter description).
If svc is set to FALSE and the data is passed locally, the argument qvec has to be specified with the column names of single-choice questions
one wants to examine. Please note that for this case the column name will be displayed as the title while the row values will be displayed as
the choice labels.
}
