% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_choice_questions.R
\name{rename_schema}
\alias{rename_schema}
\title{Changes the name of the labels and choices column in an ODK form schema.}
\usage{
rename_schema(df, lang = NULL)
}
\arguments{
\item{df}{Data frame containing the ODK form schema}

\item{lang}{Language of the choice/label column which is to be changed. Optional, defaults to NULL.}
}
\value{
Data frame
}
\description{
This function to give the column used in another function a uniform name so that no matter what language (or no language at all) is chosen,
all functions still work as they refer to the uniform name given by this function.
}
