% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ruODK.R
\name{setup_ruODK}
\alias{setup_ruODK}
\title{Connects R session to a form on ODK Central.}
\usage{
setup_ruODK(
  svc,
  un = Sys.getenv("ODKC_UN"),
  pw = Sys.getenv("ODKC_PW"),
  tz = Sys.getenv("ODKC_TZ"),
  verbose = TRUE
)
}
\arguments{
\item{svc}{Character that specifies the svc link which is used to identify an ODK form.}

\item{un}{Character that specifies the ODK Central username. Optional, defaults to Sys.getenv('ODKC_UN').}

\item{pw}{Character that specifies the ODK Central password. Optional, defaults to Sys.getenv('ODKC_PW').}

\item{tz}{Character that specifies the ODK Central tiem zone Optional, defaults to Sys.getenv('ODKC_TZ').}

\item{verbose}{Logical that sets the 'verbose' argument of the \code{\link[ruODK]{ru_setup}} function.}
}
\value{
None
}
\description{
Establishes connection with ODK Central servers by passing log in credentials, timezone and svc to the
\code{\link[ruODK]{ru_setup}} function. This is required in order to use most ruODK functions.
By default, it tries to retrieve this information from local system variables.
}
