% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission_goal_donut.R
\name{submission_goal_donut}
\alias{submission_goal_donut}
\title{Generates a pie plot (donut) that visualizes the relation between actually received and targeted number of submissions.}
\usage{
submission_goal_donut(
  df = NULL,
  csv = NULL,
  svc = FALSE,
  daily_submission_goal,
  date_col,
  exclude_weekend = TRUE
)
}
\arguments{
\item{df}{Data frame containing the ODK data that is to be used. Optional, defaults to NULL.}

\item{csv}{Character that specifies the path to the csv file that is to be read. Optional, defaults to NULL.}

\item{svc}{Logical that indicates whether the data shall be parsed using ruODK's \code{\link[ruODK]{odata_submission_get}}. Optional, defaults to FALSE.}

\item{daily_submission_goal}{Integer or float that reflects the targeted number of daily submissions.}

\item{date_col}{String that specifies the date or time stamp column in the data which is to be examined.}

\item{exclude_weekend}{Logical that determines whether weekends are excluded in the plot. Optional, defaults to TRUE.}
}
\value{
Plotly html-widget
}
\description{
The pie plot this function creates is generated by comparing the number of received submissions with the number of targeted submissions.
The target is calculated by multiplying the number-of-submissions goal per day times the number of days the data collection has been ongoing.
The number of days are determined using the function \code{\link{collection_period}}
If the targeted number of submissions has been exceeded the plot is all green and additionally provides the info by how many submissions the target number was exceeded.
Please note, that one and only one of the three data arguments (df, csv, svc) must be specified.
}
